+++ apps/teskooano/tsconfig.json
@@ -16,7 +16,7 @@
     "noUnusedParameters": true,
     "noImplicitReturns": true,
     "skipLibCheck": true,
-    "types": ["vitest/globals"]
+    "types": ["vitest/globals", "vite-plugin-glsl/ext"]
   },
   "include": ["src"]
 }
