+++ apps/teskooano/vite.d.ts
@@ -1,0 +1,2 @@
+/// <reference types="vite" />
+/// <reference types="vite-plugin-glsl/ext" />
