--- glsl.d.ts
@@ -1,8 +1,0 @@
-declare module "*.glsl" {
-  const content: string;
-  export default content;
-}
-declare module "*.glsl?raw" {
-  const content: string;
-  export default content;
-}
