--- packages/systems/celestial/src/shims-glsl.d.ts
@@ -1,19 +1,0 @@
-declare module "*.glsl" {
-  const content: string;
-  export default content;
-}
-
-declare module "*.vert" {
-  const content: string;
-  export default content;
-}
-
-declare module "*.frag" {
-  const content: string;
-  export default content;
-}
-
-declare module "*.glsl?raw" {
-  const content: string;
-  export default content;
-}
